pyInterface: Add first version of the file manager.
Add a first version of the file manager python class which should handle all
the file management of data and amplitude files (and possibly fit results and
fit configurations, but we are not there yet). The class is sparsely filled and
in an in-between state and more functionality will be added in the next
commits. Also add a small test script which tests the functionality which is
there at the moment, which will probably be deleted later on.

There was a function which was not exported in the python bindings of the
eventMetadata class. Fix this so it can be used in the file manager.