pyInterface: add reorder machanism in convertEvtToTree.py
the new reordering mechanism in the converter script still allows for the
switching of orders of the particles in different events, but keeps the order
of the first event and applies this order to the rest of the events. Thus event
files that keep their order consistent throughout the file but are not ordered
by particle ID in the first event don't get changed by the converter.