pyInterface: support loading/saving of fileManager to hard drive
create functions to save/load a fileManager object to/from the hard drive.
These functions are located in pyInterface/package/_fileManager.py and are
executed by calling pyRootPwa.core.saveFileManager (or load). Create a new
script that creates a fileManager and saves it to the hard drive for later use.
The metadata eventsType enum has to be internally handled by a python class
that resembles an enum to allow pickle to save it to the hard drive (bug in
boost::python).
Change the fileManager so that it takes a config object to initialize instead
of a config filename (config object is initialized in the outer scripts).
Remove the initialization of the particleDataTable (should also happen in outer
scripts). Do not calculate the amplitude file names with a function call each
time you need them, but calculate them once upon initialization and save them
to a list.